@@ -17,17 +17,16 @@
 
 import base64
 import copy
-import datetime
+import contextlib
 import enum
 import inspect
 import io
 import json
 import os
 import re
-from typing import Any, Literal, Optional, Union
+from typing import Any, Literal, Optional, Union, Iterator, AsyncIterator
 
 import google.auth
-from requests.exceptions import HTTPError
 
 from . import errors
 from ._api_client import BaseApiClient
@@ -56,8 +55,13 @@ def _normalize_json_case(obj: Any) -> Any:
     return [_normalize_json_case(item) for item in obj]
   elif isinstance(obj, enum.Enum):
     return obj.value
-  else:
-    return obj
+  elif isinstance(obj, str):
+    # Python >= 3.14 has a new division by zero error message.
+    if 'division by zero' in obj:
+      return obj.replace(
+          'division by zero', 'integer division or modulo by zero'
+      )
+  return obj
 
 
 def _equals_ignore_key_case(obj1: Any, obj2: Any) -> bool:
@@ -88,7 +92,7 @@ def _equals_ignore_key_case(obj1: Any, obj2: Any) -> bool:
 
 def _redact_version_numbers(version_string: str) -> str:
   """Redacts version numbers in the form x.y.z from a string."""
-  return re.sub(r'\d+\.\d+\.\d+', '{VERSION_NUMBER}', version_string)
+  return re.sub(r'\d+\.\d+\.\d+[a-zA-Z0-9]*', '{VERSION_NUMBER}', version_string)
 
 
 def _redact_language_label(language_label: str) -> str:
@@ -138,7 +142,7 @@ def _redact_request_url(url: str) -> str:
       result,
   )
   result = re.sub(
-      r'https://generativelanguage.googleapis.com/[^/]+',
+      r'.*generativelanguage.*.googleapis.com/[^/]+',
       '{MLDEV_URL_PREFIX}',
       result,
   )
@@ -204,6 +208,22 @@ def pop_undeterministic_headers(headers: dict[str, str]) -> None:
   headers.pop('Date', None)  # pytype: disable=attribute-error
   headers.pop('Server-Timing', None)  # pytype: disable=attribute-error
 
+
+@contextlib.contextmanager
+def _record_on_api_error(client: 'ReplayApiClient', http_request: HttpRequest) -> Iterator[None]:
+  try:
+    yield
+  except errors.APIError as e:
+    client._record_interaction(http_request, e)
+    raise e
+
+@contextlib.asynccontextmanager
+async def _async_record_on_api_error(client: 'ReplayApiClient', http_request: HttpRequest) -> AsyncIterator[None]:
+  try:
+    yield
+  except errors.APIError as e:
+    client._record_interaction(http_request, e)
+    raise e
 
 class ReplayRequest(BaseModel):
   """Represents a single request in a replay."""
@@ -395,6 +415,8 @@ class ReplayApiClient(BaseApiClient):
       http_request: HttpRequest,
       interaction: ReplayInteraction,
   ) -> None:
+    _debug_print(f'http_request.url: {http_request.url}')
+    _debug_print(f'interaction.request.url: {interaction.request.url}')
     assert http_request.url == interaction.request.url
     assert http_request.headers == interaction.request.headers, (
         'Request headers mismatch:\n'
@@ -467,7 +489,9 @@ class ReplayApiClient(BaseApiClient):
 
     if isinstance(response_model, list):
       response_model = response_model[0]
-    print('response_model: ', response_model.model_dump(exclude_none=True))
+    _debug_print(
+        f'response_model: {response_model.model_dump(exclude_none=True)}'
+    )
     actual = response_model.model_dump(exclude_none=True, mode='json')
     expected = interaction.response.sdk_response_segments[
         self._sdk_response_index
@@ -480,8 +504,8 @@ class ReplayApiClient(BaseApiClient):
       ):
         if 'body' in expected['sdk_http_response']:
           raw_body = expected['sdk_http_response']['body']
-          print('raw_body length: ', len(raw_body))
-          print('raw_body: ', raw_body)
+          _debug_print(f'raw_body length: {len(raw_body)}')
+          _debug_print(f'raw_body: {raw_body}')
           if isinstance(raw_body, str) and raw_body != '':
             raw_body = json.loads(raw_body)
             raw_body = json.dumps(raw_body)
@@ -491,7 +515,7 @@ class ReplayApiClient(BaseApiClient):
           actual == expected
       ), f'SDK response mismatch:\nActual: {actual}\nExpected: {expected}'
     else:
-      print('Expected SDK response mismatch:\nActual: {actual}\nExpected: {expected}')
+      _debug_print(f'Expected SDK response mismatch:\nActual: {actual}\nExpected: {expected}')
     self._sdk_response_index += 1
 
   def _request(
@@ -503,11 +527,8 @@ class ReplayApiClient(BaseApiClient):
     self._initialize_replay_session_if_not_loaded()
     if self._should_call_api():
       _debug_print('api mode request: %s' % http_request)
-      try:
+      with _record_on_api_error(self, http_request):
         result = super()._request(http_request, http_options, stream)
-      except errors.APIError as e:
-        self._record_interaction(http_request, e)
-        raise e
       if stream:
         result_segments = []
         for segment in result.segments():
@@ -532,13 +553,10 @@ class ReplayApiClient(BaseApiClient):
     self._initialize_replay_session_if_not_loaded()
     if self._should_call_api():
       _debug_print('api mode request: %s' % http_request)
-      try:
+      async with _async_record_on_api_error(self, http_request):
         result = await super()._async_request(
             http_request, http_options, stream
         )
-      except errors.APIError as e:
-        self._record_interaction(http_request, e)
-        raise e
       if stream:
         result_segments = []
         async for segment in result.async_segments():
@@ -578,16 +596,10 @@ class ReplayApiClient(BaseApiClient):
       )
     if self._should_call_api():
       result: Union[str, HttpResponse]
-      try:
+      with _record_on_api_error(self, request):
         result = super().upload_file(
             file_path, upload_url, upload_size, http_options=http_options
         )
-      except HTTPError as e:
-        result = HttpResponse(
-            dict(e.response.headers), [json.dumps({'reason': e.response.reason})]
-        )
-        result.status_code = e.response.status_code
-        raise e
       self._record_interaction(request, result)
       return result
     else:
@@ -617,16 +629,10 @@ class ReplayApiClient(BaseApiClient):
       )
     if self._should_call_api():
       result: HttpResponse
-      try:
+      async with _async_record_on_api_error(self, request):
         result = await super().async_upload_file(
             file_path, upload_url, upload_size, http_options=http_options
         )
-      except HTTPError as e:
-        result = HttpResponse(
-            dict(e.response.headers), [json.dumps({'reason': e.response.reason})]
-        )
-        result.status_code = e.response.status_code
-        raise e
       self._record_interaction(request, result)
       return result
     else:
@@ -640,14 +646,8 @@ class ReplayApiClient(BaseApiClient):
         'get', path=path, request_dict={}, http_options=http_options
     )
     if self._should_call_api():
-      try:
+      with _record_on_api_error(self, request):
         result = super().download_file(path, http_options=http_options)
-      except HTTPError as e:
-        result = HttpResponse(
-            dict(e.response.headers), [json.dumps({'reason': e.response.reason})]
-        )
-        result.status_code = e.response.status_code
-        raise e
       self._record_interaction(request, result)
       return result
     else:
@@ -661,16 +661,10 @@ class ReplayApiClient(BaseApiClient):
         'get', path=path, request_dict={}, http_options=http_options
     )
     if self._should_call_api():
-      try:
+      async with _async_record_on_api_error(self, request):
         result = await super().async_download_file(
             path, http_options=http_options
         )
-      except HTTPError as e:
-        result = HttpResponse(
-            dict(e.response.headers), [json.dumps({'reason': e.response.reason})]
-        )
-        result.status_code = e.response.status_code
-        raise e
       self._record_interaction(request, result)
       return result
     else:
