@@ -27,6 +27,7 @@ from . import _transformers as t
 from . import types
 from ._api_client import BaseApiClient
 from ._common import get_value_by_path as getv
+from ._common import move_value_by_path as movev
 from ._common import set_value_by_path as setv
 from .pagers import AsyncPager, Pager
 
@@ -327,6 +328,11 @@ def _BatchJob_from_vertex(
         ),
     )
 
+  if getv(from_object, ['completionStats']) is not None:
+    setv(
+        to_object, ['completion_stats'], getv(from_object, ['completionStats'])
+    )
+
   return to_object
 
 
@@ -335,11 +341,11 @@ def _Blob_to_mldev(
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['display_name']) is not None:
-    raise ValueError('display_name parameter is not supported in Gemini API.')
-
   if getv(from_object, ['data']) is not None:
     setv(to_object, ['data'], getv(from_object, ['data']))
+
+  if getv(from_object, ['display_name']) is not None:
+    raise ValueError('display_name parameter is not supported in Gemini API.')
 
   if getv(from_object, ['mime_type']) is not None:
     setv(to_object, ['mimeType'], getv(from_object, ['mime_type']))
@@ -402,13 +408,6 @@ def _Candidate_from_mldev(
   if getv(from_object, ['finishReason']) is not None:
     setv(to_object, ['finish_reason'], getv(from_object, ['finishReason']))
 
-  if getv(from_object, ['urlContextMetadata']) is not None:
-    setv(
-        to_object,
-        ['url_context_metadata'],
-        getv(from_object, ['urlContextMetadata']),
-    )
-
   if getv(from_object, ['avgLogprobs']) is not None:
     setv(to_object, ['avg_logprobs'], getv(from_object, ['avgLogprobs']))
 
@@ -430,6 +429,13 @@ def _Candidate_from_mldev(
         to_object,
         ['safety_ratings'],
         [item for item in getv(from_object, ['safetyRatings'])],
+    )
+
+  if getv(from_object, ['urlContextMetadata']) is not None:
+    setv(
+        to_object,
+        ['url_context_metadata'],
+        getv(from_object, ['urlContextMetadata']),
     )
 
   return to_object
@@ -713,10 +719,10 @@ def _EmbedContentBatch_to_mldev(
   if getv(from_object, ['config']) is not None:
     setv(
         to_object,
-        ['config'],
+        ['_self'],
         _EmbedContentConfig_to_mldev(getv(from_object, ['config']), to_object),
     )
-
+    movev(to_object, {'requests[].*': 'requests[].request.*'})
   return to_object
 
 
@@ -1015,19 +1021,38 @@ def _GetBatchJobParameters_to_vertex(
   return to_object
 
 
+def _GoogleMaps_to_mldev(
+    from_object: Union[dict[str, Any], object],
+    parent_object: Optional[dict[str, Any]] = None,
+) -> dict[str, Any]:
+  to_object: dict[str, Any] = {}
+  if getv(from_object, ['auth_config']) is not None:
+    raise ValueError('auth_config parameter is not supported in Gemini API.')
+
+  if getv(from_object, ['enable_widget']) is not None:
+    setv(to_object, ['enableWidget'], getv(from_object, ['enable_widget']))
+
+  return to_object
+
+
 def _GoogleSearch_to_mldev(
     from_object: Union[dict[str, Any], object],
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['time_range_filter']) is not None:
-    setv(
-        to_object, ['timeRangeFilter'], getv(from_object, ['time_range_filter'])
-    )
-
   if getv(from_object, ['exclude_domains']) is not None:
     raise ValueError(
         'exclude_domains parameter is not supported in Gemini API.'
+    )
+
+  if getv(from_object, ['blocking_confidence']) is not None:
+    raise ValueError(
+        'blocking_confidence parameter is not supported in Gemini API.'
+    )
+
+  if getv(from_object, ['time_range_filter']) is not None:
+    setv(
+        to_object, ['timeRangeFilter'], getv(from_object, ['time_range_filter'])
     )
 
   return to_object
@@ -1056,12 +1081,17 @@ def _InlinedRequest_to_mldev(
         ],
     )
 
+  if getv(from_object, ['metadata']) is not None:
+    setv(to_object, ['metadata'], getv(from_object, ['metadata']))
+
   if getv(from_object, ['config']) is not None:
     setv(
         to_object,
         ['request', 'generationConfig'],
         _GenerateContentConfig_to_mldev(
-            api_client, getv(from_object, ['config']), to_object
+            api_client,
+            getv(from_object, ['config']),
+            getv(to_object, ['request'], default_value={}),
         ),
     )
 
@@ -1215,33 +1245,6 @@ def _Part_to_mldev(
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['video_metadata']) is not None:
-    setv(to_object, ['videoMetadata'], getv(from_object, ['video_metadata']))
-
-  if getv(from_object, ['thought']) is not None:
-    setv(to_object, ['thought'], getv(from_object, ['thought']))
-
-  if getv(from_object, ['inline_data']) is not None:
-    setv(
-        to_object,
-        ['inlineData'],
-        _Blob_to_mldev(getv(from_object, ['inline_data']), to_object),
-    )
-
-  if getv(from_object, ['file_data']) is not None:
-    setv(
-        to_object,
-        ['fileData'],
-        _FileData_to_mldev(getv(from_object, ['file_data']), to_object),
-    )
-
-  if getv(from_object, ['thought_signature']) is not None:
-    setv(
-        to_object,
-        ['thoughtSignature'],
-        getv(from_object, ['thought_signature']),
-    )
-
   if getv(from_object, ['function_call']) is not None:
     setv(to_object, ['functionCall'], getv(from_object, ['function_call']))
 
@@ -1255,6 +1258,13 @@ def _Part_to_mldev(
   if getv(from_object, ['executable_code']) is not None:
     setv(to_object, ['executableCode'], getv(from_object, ['executable_code']))
 
+  if getv(from_object, ['file_data']) is not None:
+    setv(
+        to_object,
+        ['fileData'],
+        _FileData_to_mldev(getv(from_object, ['file_data']), to_object),
+    )
+
   if getv(from_object, ['function_response']) is not None:
     setv(
         to_object,
@@ -1262,8 +1272,28 @@ def _Part_to_mldev(
         getv(from_object, ['function_response']),
     )
 
+  if getv(from_object, ['inline_data']) is not None:
+    setv(
+        to_object,
+        ['inlineData'],
+        _Blob_to_mldev(getv(from_object, ['inline_data']), to_object),
+    )
+
   if getv(from_object, ['text']) is not None:
     setv(to_object, ['text'], getv(from_object, ['text']))
+
+  if getv(from_object, ['thought']) is not None:
+    setv(to_object, ['thought'], getv(from_object, ['thought']))
+
+  if getv(from_object, ['thought_signature']) is not None:
+    setv(
+        to_object,
+        ['thoughtSignature'],
+        getv(from_object, ['thought_signature']),
+    )
+
+  if getv(from_object, ['video_metadata']) is not None:
+    setv(to_object, ['videoMetadata'], getv(from_object, ['video_metadata']))
 
   return to_object
 
@@ -1273,11 +1303,11 @@ def _SafetySetting_to_mldev(
     parent_object: Optional[dict[str, Any]] = None,
 ) -> dict[str, Any]:
   to_object: dict[str, Any] = {}
-  if getv(from_object, ['method']) is not None:
-    raise ValueError('method parameter is not supported in Gemini API.')
-
   if getv(from_object, ['category']) is not None:
     setv(to_object, ['category'], getv(from_object, ['category']))
+
+  if getv(from_object, ['method']) is not None:
+    raise ValueError('method parameter is not supported in Gemini API.')
 
   if getv(from_object, ['threshold']) is not None:
     setv(to_object, ['threshold'], getv(from_object, ['threshold']))
@@ -1300,13 +1330,6 @@ def _Tool_to_mldev(
   if getv(from_object, ['retrieval']) is not None:
     raise ValueError('retrieval parameter is not supported in Gemini API.')
 
-  if getv(from_object, ['google_search']) is not None:
-    setv(
-        to_object,
-        ['googleSearch'],
-        _GoogleSearch_to_mldev(getv(from_object, ['google_search']), to_object),
-    )
-
   if getv(from_object, ['google_search_retrieval']) is not None:
     setv(
         to_object,
@@ -1314,22 +1337,36 @@ def _Tool_to_mldev(
         getv(from_object, ['google_search_retrieval']),
     )
 
+  if getv(from_object, ['computer_use']) is not None:
+    setv(to_object, ['computerUse'], getv(from_object, ['computer_use']))
+
+  if getv(from_object, ['file_search']) is not None:
+    setv(to_object, ['fileSearch'], getv(from_object, ['file_search']))
+
+  if getv(from_object, ['code_execution']) is not None:
+    setv(to_object, ['codeExecution'], getv(from_object, ['code_execution']))
+
   if getv(from_object, ['enterprise_web_search']) is not None:
     raise ValueError(
         'enterprise_web_search parameter is not supported in Gemini API.'
     )
 
   if getv(from_object, ['google_maps']) is not None:
-    raise ValueError('google_maps parameter is not supported in Gemini API.')
+    setv(
+        to_object,
+        ['googleMaps'],
+        _GoogleMaps_to_mldev(getv(from_object, ['google_maps']), to_object),
+    )
+
+  if getv(from_object, ['google_search']) is not None:
+    setv(
+        to_object,
+        ['googleSearch'],
+        _GoogleSearch_to_mldev(getv(from_object, ['google_search']), to_object),
+    )
 
   if getv(from_object, ['url_context']) is not None:
     setv(to_object, ['urlContext'], getv(from_object, ['url_context']))
-
-  if getv(from_object, ['computer_use']) is not None:
-    setv(to_object, ['computerUse'], getv(from_object, ['computer_use']))
-
-  if getv(from_object, ['code_execution']) is not None:
-    setv(to_object, ['codeExecution'], getv(from_object, ['code_execution']))
 
   return to_object
 
@@ -1794,36 +1831,11 @@ class Batches(_api_module.BaseModule):
         config=config,
     )
 
-    http_options: Optional[types.HttpOptions] = None
-    if (
-        parameter_model.config is not None
-        and parameter_model.config.http_options is not None
-    ):
-      http_options = parameter_model.config.http_options
-
     if self._api_client.vertexai:
       config = _extra_utils.format_destination(src, parameter_model.config)
       return self._create(model=model, src=src, config=config)
-    elif src.inlined_requests is None:
+    else:
       return self._create(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_generate_content_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = self._api_client.request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   def create_embeddings(
       self,
@@ -1870,35 +1882,10 @@ class Batches(_api_module.BaseModule):
         config=config,
     )
 
-    http_options: Optional[types.HttpOptions] = None
-    if (
-        parameter_model.config is not None
-        and parameter_model.config.http_options is not None
-    ):
-      http_options = parameter_model.config.http_options
-
     if self._api_client.vertexai:
       raise ValueError('Vertex AI does not support batches.create_embeddings.')
-    elif src.inlined_requests is None:
+    else:
       return self._create_embeddings(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_embedding_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = self._api_client.request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   def list(
       self, *, config: Optional[types.ListBatchJobsConfigOrDict] = None
@@ -2393,36 +2380,11 @@ class AsyncBatches(_api_module.BaseModule):
         config=config,
     )
 
-    http_options: Optional[types.HttpOptions] = None
-    if (
-        parameter_model.config is not None
-        and parameter_model.config.http_options is not None
-    ):
-      http_options = parameter_model.config.http_options
-
     if self._api_client.vertexai:
       config = _extra_utils.format_destination(src, parameter_model.config)
       return await self._create(model=model, src=src, config=config)
-    elif src.inlined_requests is None:
+    else:
       return await self._create(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_generate_content_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = await self._api_client.async_request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   async def create_embeddings(
       self,
@@ -2478,26 +2440,8 @@ class AsyncBatches(_api_module.BaseModule):
 
     if self._api_client.vertexai:
       raise ValueError('Vertex AI does not support batches.create_embeddings.')
-    elif src.inlined_requests is None:
+    else:
       return await self._create_embeddings(model=model, src=src, config=config)
-
-    path, request_dict = _create_inlined_embedding_request_dict(
-        self._api_client, parameter_model
-    )
-
-    response = await self._api_client.async_request(
-        'post', path, request_dict, http_options
-    )
-
-    response_dict = '' if not response.body else json.loads(response.body)
-    response_dict = _BatchJob_from_mldev(response_dict)
-
-    return_value = types.BatchJob._from_response(
-        response=response_dict, kwargs=parameter_model.model_dump()
-    )
-
-    self._api_client._verify_response(return_value)
-    return return_value
 
   async def list(
       self, *, config: Optional[types.ListBatchJobsConfigOrDict] = None
@@ -2528,100 +2472,3 @@ class AsyncBatches(_api_module.BaseModule):
         await self._list(config=config),
         config,
     )
-
-
-def _create_inlined_generate_content_request_dict(
-    client: BaseApiClient, parameter_model: types._CreateBatchJobParameters
-) -> tuple[str, dict[str, Any]]:
-  request_url_dict: Optional[dict[str, str]]
-
-  request_dict: dict[str, Any] = _CreateBatchJobParameters_to_mldev(
-      client, parameter_model
-  )
-
-  request_url_dict = request_dict.get('_url')
-  if request_url_dict:
-    path = '{model}:batchGenerateContent'.format_map(request_url_dict)
-  else:
-    path = '{model}:batchGenerateContent'
-  query_params = request_dict.get('_query')
-  if query_params:
-    path = f'{path}?{urlencode(query_params)}'
-  request_dict.pop('config', None)
-
-  request_dict = _common.convert_to_dict(request_dict)
-  request_dict = _common.encode_unserializable_types(request_dict)
-  # Move system instruction to 'request':
-  # {'systemInstruction': system_instruction}
-  requests = []
-  batch_dict = request_dict.get('batch')
-  if batch_dict and isinstance(batch_dict, dict):
-    input_config_dict = batch_dict.get('inputConfig')
-    if input_config_dict and isinstance(input_config_dict, dict):
-      requests_dict = input_config_dict.get('requests')
-      if requests_dict and isinstance(requests_dict, dict):
-        requests = requests_dict.get('requests')
-  new_requests = []
-  if requests:
-    for req in requests:
-      if req.get('systemInstruction'):
-        value = req.pop('systemInstruction')
-        req['request'].update({'systemInstruction': value})
-      new_requests.append(req)
-  request_dict['batch']['inputConfig']['requests'][  # type: ignore
-      'requests'
-  ] = new_requests
-  return path, request_dict
-
-
-def _create_inlined_embedding_request_dict(
-    client: BaseApiClient,
-    parameter_model: types._CreateEmbeddingsBatchJobParameters,
-) -> tuple[str, dict[str, Any]]:
-  src = parameter_model.src
-  if not isinstance(src, types.EmbeddingsBatchJobSource):
-    raise ValueError(f'Invalid batch job source: {src}.')
-
-  request_url_dict: Optional[dict[str, str]]
-
-  request_dict: dict[str, Any] = _CreateEmbeddingsBatchJobParameters_to_mldev(
-      client, parameter_model
-  )
-
-  request_url_dict = request_dict.get('_url')
-  if request_url_dict:
-    path = '{model}:asyncBatchEmbedContent'.format_map(request_url_dict)
-  else:
-    path = '{model}:asyncBatchEmbedContent'
-  query_params = request_dict.get('_query')
-  if query_params:
-    path = f'{path}?{urlencode(query_params)}'
-
-  request_dict.pop('config', None)
-  request_dict.get('batch', {}).get('inputConfig', {}).get('requests', {}).pop(
-      'config', None
-  )
-
-  request_dict = _common.convert_to_dict(request_dict)
-  request_dict = _common.encode_unserializable_types(request_dict)
-
-  requests = []
-  batch_dict = request_dict.get('batch')
-  if batch_dict and isinstance(batch_dict, dict):
-    input_config_dict = batch_dict.get('inputConfig')
-    if input_config_dict and isinstance(input_config_dict, dict):
-      requests_dict = input_config_dict.get('requests')
-      if requests_dict and isinstance(requests_dict, dict):
-        requests = requests_dict.get('requests')
-  new_requests = []
-  if requests:
-    for req in requests:
-      for k in list(req.keys()):
-        if k != 'request':
-          req['request'][k] = req.pop(k)
-      new_requests.append(req)
-  request_dict['batch']['inputConfig']['requests'][  # type: ignore
-      'requests'
-  ] = new_requests
-
-  return path, request_dict
