@@ -27,6 +27,7 @@ from ._replay_api_client import ReplayApiClient
 from .batches import AsyncBatches, Batches
 from .caches import AsyncCaches, Caches
 from .chats import AsyncChats, Chats
+from .file_search_stores import AsyncFileSearchStores, FileSearchStores
 from .files import AsyncFiles, Files
 from .live import AsyncLive
 from .models import AsyncModels, Models
@@ -47,6 +48,7 @@ class AsyncClient:
     self._caches = AsyncCaches(self._api_client)
     self._batches = AsyncBatches(self._api_client)
     self._files = AsyncFiles(self._api_client)
+    self._file_search_stores = AsyncFileSearchStores(self._api_client)
     self._live = AsyncLive(self._api_client)
     self._tokens = AsyncTokens(self._api_client)
     self._operations = AsyncOperations(self._api_client)
@@ -62,6 +64,10 @@ class AsyncClient:
   @property
   def caches(self) -> AsyncCaches:
     return self._caches
+
+  @property
+  def file_search_stores(self) -> AsyncFileSearchStores:
+    return self._file_search_stores
 
   @property
   def batches(self) -> AsyncBatches:
@@ -276,6 +282,7 @@ class Client:
     self._models = Models(self._api_client)
     self._tunings = Tunings(self._api_client)
     self._caches = Caches(self._api_client)
+    self._file_search_stores = FileSearchStores(self._api_client)
     self._batches = Batches(self._api_client)
     self._files = Files(self._api_client)
     self._tokens = Tokens(self._api_client)
@@ -338,6 +345,10 @@ class Client:
     return self._caches
 
   @property
+  def file_search_stores(self) -> FileSearchStores:
+    return self._file_search_stores
+
+  @property
   def batches(self) -> Batches:
     return self._batches
 
@@ -397,4 +408,7 @@ class Client:
     self.close()
 
   def __del__(self) -> None:
-    self.close()
+    try:
+      self.close()
+    except Exception:
+      pass
