@@ -25,7 +25,13 @@ from . import types
 T = TypeVar('T')
 
 PagedItem = Literal[
-    'batch_jobs', 'models', 'tuning_jobs', 'files', 'cached_contents'
+    'batch_jobs',
+    'models',
+    'tuning_jobs',
+    'files',
+    'cached_contents',
+    'file_search_stores',
+    'documents',
 ]
 
 
