@@ -37,7 +37,7 @@ class TestLocalTokenizer(unittest.TestCase):
     self.mock_tokenizer = MagicMock()
     self.mock_get_sentencepiece.return_value = self.mock_tokenizer
 
-    self.tokenizer = local_tokenizer.LocalTokenizer(model_name='gemini-1.0-pro')
+    self.tokenizer = local_tokenizer.LocalTokenizer(model_name='gemini-2.5-flash')
 
   def tearDown(self):
     patch.stopall()
