@@ -15,6 +15,7 @@
 
 import collections
 import logging
+import os
 import sys
 import typing
 
@@ -25,6 +26,12 @@ from ... import _transformers as t
 from ... import errors
 from ... import types
 from .. import pytest_helper
+
+GOOGLE_HOMEPAGE_FILE_PATH = os.path.abspath(
+    os.path.join(os.path.dirname(__file__), '../data/google_homepage.png')
+)
+with open(GOOGLE_HOMEPAGE_FILE_PATH, 'rb') as image_file:
+  google_homepage_screenshot_bytes = image_file.read()
 
 function_declarations = [{
     'name': 'get_current_weather',
@@ -39,6 +46,31 @@ function_declarations = [{
             'unit': {
                 'type': 'STRING',
                 'enum': ['C', 'F'],
+            },
+        },
+    },
+}]
+computer_use_override_function_declarations = [{
+    'name': 'type_text_at',
+    'description': 'Types text at a certain coordinate.',
+    'parameters': {
+        'type': 'OBJECT',
+        'properties': {
+            'y': {
+                'type': 'INTEGER',
+                'description': 'The y-coordinate, normalized from 0 to 1000.',
+            },
+            'x': {
+                'type': 'INTEGER',
+                'description': 'The x-coordinate, normalized from 0 to 1000.',
+            },
+            'press_enter': {
+                'type': 'BOOLEAN',
+                'description': 'Whether to press enter after typing the text.'
+            },
+            'text': {
+                'type': 'STRING',
+                'description': 'The text to type.',
             },
         },
     },
@@ -66,6 +98,33 @@ manual_function_calling_contents = [
         }],
     },
     {'role': 'user', 'parts': function_response_parts},
+]
+computer_use_multi_turn_contents = [
+    {
+        'role': 'user',
+        'parts': [{'text': 'Go to google and search nano banana'}],
+    },
+    {
+        'role': 'model',
+        'parts': [{'function_call': {'name': 'open_web_browser', 'args': {}}}],
+    },
+    {
+        'role': 'user',
+        'parts': [{
+            'function_response': {
+                'name': 'open_web_browser',
+                'response': {
+                    'url': 'http://www.google.com',
+                },
+                'parts': [{
+                    'inline_data': {
+                        'data': google_homepage_screenshot_bytes,
+                        'mime_type': 'image/png',
+                    }
+                }],
+            }
+        }],
+    },
 ]
 
 
@@ -106,17 +165,17 @@ def divide_floats(a: float, b: float) -> float:
 
 test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
-        name='test_google_search_retrieval',
+        name='test_google_search',
         parameters=types._GenerateContentParameters(
-            model='gemini-1.5-flash',
+            model='gemini-2.5-flash',
             contents=t.t_contents('Why is the sky blue?'),
-            config={'tools': [{'google_search_retrieval': {}}]},
+            config={'tools': [{'google_search': {}}]},
         ),
     ),
     pytest_helper.TestTableItem(
         name='test_vai_search',
         parameters=types._GenerateContentParameters(
-            model='gemini-1.5-flash',
+            model='gemini-2.5-flash',
             contents=t.t_contents('what is vertex ai search?'),
             config={
                 'tools': [{
@@ -135,7 +194,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_vai_google_search',
         parameters=types._GenerateContentParameters(
-            model='gemini-1.5-flash',
+            model='gemini-2.5-flash',
             contents=t.t_contents('why is the sky blue?'),
             config={
                 'tools': [
@@ -175,7 +234,7 @@ test_table: list[pytest_helper.TestTableItem] = [
     pytest_helper.TestTableItem(
         name='test_rag_model_old',
         parameters=types._GenerateContentParameters(
-            model='gemini-1.5-flash',
+            model='gemini-2.5-flash',
             contents=t.t_contents(
                 'How much gain or loss did Google get in the Motorola Mobile'
                 ' deal in 2014?',
@@ -232,9 +291,97 @@ test_table: list[pytest_helper.TestTableItem] = [
         exception_if_mldev='retrieval',
     ),
     pytest_helper.TestTableItem(
+        name='test_file_search',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents(
+                'can you tell me the author of "A Survey of Modernist Poetry"?',
+            ),
+            config={
+                'tools': [
+                    types.Tool(
+                        file_search=types.FileSearch(
+                            file_search_store_names=[
+                                'fileSearchStores/5en07ei3kojo-yo8sjqgvx2xf'
+                            ]
+                        ),
+                    ),
+                ],
+            },
+        ),
+        exception_if_vertex='is not supported in Vertex AI',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_file_search_non_existent_file_search_store',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents(
+                'can you tell me the author of "A Survey of Modernist Poetry"?',
+            ),
+            config={
+                'tools': [
+                    types.Tool(
+                        file_search=types.FileSearch(
+                            file_search_store_names=[
+                                'fileSearchStores/test-non-existent-rag-store'
+                            ],
+                        ),
+                    ),
+                ],
+            },
+        ),
+        exception_if_mldev='not exist',
+        exception_if_vertex='is not supported in Vertex AI',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_file_search_with_metadata_filter',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents(
+                'can you tell me the author of "A Survey of Modernist Poetry"?',
+            ),
+            config={
+                'tools': [
+                    types.Tool(
+                        file_search=types.FileSearch(
+                            file_search_store_names=[
+                                'fileSearchStores/5en07ei3kojo-yo8sjqgvx2xf'
+                            ],
+                            metadata_filter='tag=science',
+                        ),
+                    ),
+                ],
+            },
+        ),
+        exception_if_vertex='is not supported in Vertex AI',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_file_search_with_metadata_filter_and_top_k',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents(
+                'can you tell me the author of "A Survey of Modernist Poetry"',
+            ),
+            config={
+                'tools': [
+                    types.Tool(
+                        file_search=types.FileSearch(
+                            file_search_store_names=[
+                                'fileSearchStores/5en07ei3kojo-yo8sjqgvx2xf'
+                            ],
+                            metadata_filter='tag=science',
+                            top_k=1,
+                        ),
+                    ),
+                ],
+            },
+        ),
+        exception_if_vertex='is not supported in Vertex AI',
+    ),
+    pytest_helper.TestTableItem(
         name='test_function_call',
         parameters=types._GenerateContentParameters(
-            model='gemini-1.5-flash',
+            model='gemini-2.5-flash',
             contents=manual_function_calling_contents,
             config={
                 'tools': [{'function_declarations': function_declarations}]
@@ -245,12 +392,12 @@ test_table: list[pytest_helper.TestTableItem] = [
         # TODO(b/382547236) add the test back in api mode when the code
         # execution is supported.
         skip_in_api_mode=(
-            'Model gemini-1.5-flash-001 does not support code execution for'
+            'Model gemini-2.5-flash-001 does not support code execution for'
             ' Vertex API.'
         ),
         name='test_code_execution',
         parameters=types._GenerateContentParameters(
-            model='gemini-1.5-flash',
+            model='gemini-2.5-flash',
             contents=t.t_contents(
                 'What is the sum of the first 50 prime numbers? '
                 + 'Generate and run code for the calculation, and make sure you'
@@ -260,18 +407,14 @@ test_table: list[pytest_helper.TestTableItem] = [
         ),
     ),
     pytest_helper.TestTableItem(
-        name='test_function_google_search_retrieval_with_long_lat',
+        name='test_function_google_search_with_long_lat',
         parameters=types._GenerateContentParameters(
-            model='gemini-1.5-flash',
+            model='gemini-2.5-flash',
             contents=t.t_contents('what is the price of GOOG?'),
             config=types.GenerateContentConfig(
                 tools=[
                     types.Tool(
-                        google_search_retrieval=types.GoogleSearchRetrieval(
-                            dynamic_retrieval_config=types.DynamicRetrievalConfig(
-                                mode='MODE_UNSPECIFIED'
-                            )
-                        )
+                        google_search=types.GoogleSearch(),
                     ),
                 ],
                 tool_config=types.ToolConfig(
@@ -316,6 +459,82 @@ test_table: list[pytest_helper.TestTableItem] = [
         ),
     ),
     pytest_helper.TestTableItem(
+        name='test_computer_use',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-computer-use-preview-10-2025',
+            contents=t.t_contents('Go to google and search nano banana'),
+            config={'tools': [{'computer_use': {}}]},
+        ),
+        exception_if_vertex='404',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_computer_use_with_browser_environment',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-computer-use-preview-10-2025',
+            contents=t.t_contents('Go to google and search nano banana'),
+            config={
+                'tools': [
+                    {'computer_use': {'environment': 'ENVIRONMENT_BROWSER'}}
+                ]
+            },
+        ),
+        exception_if_vertex='404',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_computer_use_multi_turn',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-computer-use-preview-10-2025',
+            contents=computer_use_multi_turn_contents,
+            config={
+                'tools': [
+                    {'computer_use': {'environment': 'ENVIRONMENT_BROWSER'}}
+                ]
+            },
+        ),
+        exception_if_vertex='404',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_computer_use_exclude_predefined_functions',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-computer-use-preview-10-2025',
+            contents='cheapest flight to NYC on Mar 18 2025 on Google Flights',
+            config={
+                'tools': [
+                    {
+                        'computer_use': {
+                            'environment': 'ENVIRONMENT_BROWSER',
+                            'excluded_predefined_functions': ['click_at'],
+                        },
+                    },
+                ]
+            },
+        ),
+        exception_if_vertex='404',
+    ),
+    pytest_helper.TestTableItem(
+        name='test_computer_use_override_default_function',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-computer-use-preview-10-2025',
+            contents=computer_use_multi_turn_contents,
+            config={
+                'tools': [
+                    {
+                        'computer_use': {
+                            'environment': 'ENVIRONMENT_BROWSER',
+                            'excluded_predefined_functions': ['type_text_at'],
+                        },
+                    },
+                    {
+                        'function_declarations': (
+                            computer_use_override_function_declarations
+                        )
+                    },
+                ]
+            },
+        ),
+        exception_if_vertex='404',
+    ),
+    pytest_helper.TestTableItem(
         # https://github.com/googleapis/python-genai/issues/830
         # - models started returning empty thought in response to queries
         #   containing tools.
@@ -328,7 +547,7 @@ test_table: list[pytest_helper.TestTableItem] = [
         #   dropped by the SDK leaving a `{'thought: True}` part.
         name='test_chat_tools_empty_thoughts',
         parameters=types._GenerateContentParameters(
-            model='gemini-2.5-flash-preview-05-20',
+            model='gemini-2.5-flash',
             contents=[
                 types.Content.model_validate(item)
                 for item in [
@@ -367,15 +586,21 @@ test_table: list[pytest_helper.TestTableItem] = [
         name='test_function_calling_config_validated_mode',
         parameters=types._GenerateContentParameters(
             model='gemini-2.5-flash',
-            contents=t.t_contents(
-                'How is the weather in Kirkland?'
-            ),
+            contents=t.t_contents('How is the weather in Kirkland?'),
             config={
                 'tools': [{'function_declarations': function_declarations}],
                 'tool_config': {
                     'function_calling_config': {'mode': 'VALIDATED'}
                 },
             },
+        ),
+    ),
+    pytest_helper.TestTableItem(
+        name='test_google_maps_with_enable_widget',
+        parameters=types._GenerateContentParameters(
+            model='gemini-2.5-flash',
+            contents=t.t_contents('What is the nearest airport to Seattle?'),
+            config={'tools': [{'google_maps': {'enable_widget': True}}]},
         ),
     ),
 ]
@@ -391,16 +616,12 @@ pytest_plugins = ('pytest_asyncio',)
 
 
 # Cannot be included in test_table because json serialization fails on function.
-def test_function_google_search_retrieval(client):
+def test_function_google_search(client):
   contents = 'What is the price of GOOG?.'
   config = types.GenerateContentConfig(
       tools=[
           types.Tool(
-              google_search_retrieval=types.GoogleSearchRetrieval(
-                  dynamic_retrieval_config=types.DynamicRetrievalConfig(
-                      mode='MODE_UNSPECIFIED'
-                  )
-              )
+              google_search=types.GoogleSearch(),
           ),
           get_stock_price,
       ],
@@ -411,21 +632,21 @@ def test_function_google_search_retrieval(client):
   # bad request to combine function call and google search retrieval
   with pytest.raises(errors.ClientError):
     client.models.generate_content(
-        model='gemini-1.5-flash',
+        model='gemini-2.5-flash',
         contents=contents,
         config=config,
     )
 
 
-def test_google_search_retrieval_stream(client):
+def test_google_search_stream(client):
   for part in client.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents=types.Content(
           role='user',
           parts=[types.Part(text='Why is the sky blue?')],
       ),
       config=types.GenerateContentConfig(
-          tools=[types.ToolDict({'google_search_retrieval': {}})],
+          tools=[types.ToolDict({'google_search': {}})],
       ),
   ):
     pass
@@ -440,7 +661,7 @@ def test_google_search_retrieval_stream(client):
 )
 def test_function_calling_without_implementation(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='What is the weather in Boston?',
       config={
           'tools': [get_weather_declaration_only],
@@ -451,7 +672,7 @@ def test_function_calling_without_implementation(client):
 
 def test_2_function(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='What is the price of GOOG? And what is the weather in Boston?',
       config={
           'tools': [get_weather, get_stock_price],
@@ -466,7 +687,7 @@ def test_2_function(client):
 @pytest.mark.asyncio
 async def test_2_function_async(client):
   response = await client.aio.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='What is the price of GOOG? And what is the weather in Boston?',
       config={
           'tools': [get_weather, get_stock_price],
@@ -495,7 +716,7 @@ def test_automatic_function_calling_with_customized_math_rule(client):
 
 def test_automatic_function_calling(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -509,7 +730,7 @@ def test_automatic_function_calling(client):
 @pytest.mark.asyncio
 async def test_automatic_function_calling_with_async_function(client):
   response = await client.aio.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1001.0/2.0?',
       config={
           'tools': [divide_floats_async],
@@ -522,7 +743,7 @@ async def test_automatic_function_calling_with_async_function(client):
 
 def test_automatic_function_calling_stream(client):
   response = client.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -538,7 +759,7 @@ def test_automatic_function_calling_stream(client):
 def test_disable_automatic_function_calling_stream(client):
   # If AFC is disabled, the response should contain a function call.
   response = client.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -553,7 +774,7 @@ def test_disable_automatic_function_calling_stream(client):
 
 def test_automatic_function_calling_no_function_response_stream(client):
   response = client.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the weather in Boston?',
       config={
           'tools': [divide_integers],
@@ -570,7 +791,7 @@ def test_automatic_function_calling_no_function_response_stream(client):
 async def test_disable_automatic_function_calling_stream_async(client):
   # If AFC is disabled, the response should contain a function call.
   response = await client.aio.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -588,7 +809,7 @@ async def test_automatic_function_calling_no_function_response_stream_async(
     client,
 ):
   response = await client.aio.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the weather in Boston?',
       config={
           'tools': [divide_integers],
@@ -604,7 +825,7 @@ async def test_automatic_function_calling_no_function_response_stream_async(
 @pytest.mark.asyncio
 async def test_automatic_function_calling_stream_async(client):
   response = await client.aio.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -619,7 +840,7 @@ async def test_automatic_function_calling_stream_async(client):
 
 def test_callable_tools_user_disable_afc(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -633,7 +854,7 @@ def test_callable_tools_user_disable_afc(client):
 
 def test_callable_tools_user_disable_afc_with_max_remote_calls(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -650,7 +871,7 @@ def test_callable_tools_user_disable_afc_with_max_remote_calls_negative(
     client,
 ):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -665,7 +886,7 @@ def test_callable_tools_user_disable_afc_with_max_remote_calls_negative(
 
 def test_callable_tools_user_disable_afc_with_max_remote_calls_zero(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -680,7 +901,7 @@ def test_callable_tools_user_disable_afc_with_max_remote_calls_zero(client):
 
 def test_callable_tools_user_enable_afc(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -694,7 +915,7 @@ def test_callable_tools_user_enable_afc(client):
 
 def test_callable_tools_user_enable_afc_with_max_remote_calls(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -711,7 +932,7 @@ def test_callable_tools_user_enable_afc_with_max_remote_calls_negative(
     client,
 ):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -726,7 +947,7 @@ def test_callable_tools_user_enable_afc_with_max_remote_calls_negative(
 
 def test_callable_tools_user_enable_afc_with_max_remote_calls_zero(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -741,7 +962,7 @@ def test_callable_tools_user_enable_afc_with_max_remote_calls_zero(client):
 
 def test_automatic_function_calling_with_exception(client):
   client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/0?',
       config={
           'tools': [divide_integers],
@@ -752,7 +973,7 @@ def test_automatic_function_calling_with_exception(client):
 
 def test_automatic_function_calling_float_without_decimal(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000.0/2.0?',
       config={
           'tools': [divide_floats, divide_integers],
@@ -776,7 +997,7 @@ def test_automatic_function_calling_with_pydantic_model(client):
       return f'The weather in {city_object.city_name} is sunny and 100 degrees.'
 
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='it is winter now, what is the weather in Boston?',
       config={
           'tools': [get_weather_pydantic_model],
@@ -809,7 +1030,7 @@ def test_automatic_function_calling_with_pydantic_model_in_list_type(client):
     return result
 
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='it is winter now, what is the weather in Boston and New York?',
       config={
           'tools': [get_weather_from_list_of_cities],
@@ -850,14 +1071,17 @@ def test_automatic_function_calling_with_pydantic_model_in_union_type(client):
     else:
       return 'The animal is not supported'
 
-  with pytest_helper.exception_if_mldev(client, errors.ClientError):
+  with pytest_helper.exception_if_vertex(client, errors.ClientError):
     response = client.models.generate_content(
-        model='gemini-1.5-flash',
+        model='gemini-2.5-flash',
         contents=(
             'I have a one year old cat named Sundae, can you get the'
             ' information of the cat for me?'
         ),
         config={
+            'system_instruction': (
+                'you answer questions based on the tools provided'
+            ),
             'tools': [get_information],
             'automatic_function_calling': {'ignore_call_history': True},
         },
@@ -884,7 +1108,7 @@ def test_automatic_function_calling_with_union_operator(client):
       return f'The object of interest is {object_of_interest}'
 
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents=(
           'I have a one year old cat named Sundae, can you get the'
           ' information of the cat for me?'
@@ -904,7 +1128,7 @@ def test_automatic_function_calling_with_tuple_param(client):
     return f'The latitude is {latlng[0]} and the longitude is {latlng[1]}'
 
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents=(
           'The coordinates are (51.509, -0.118). What is the latitude and longitude?'
       ),
@@ -965,7 +1189,7 @@ def test_automatic_function_calling_with_parameterized_generic_union_type(
       )
 
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='Can you describe the city of San Francisco, USA?',
       config={
           'tools': [describe_cities],
@@ -976,21 +1200,21 @@ def test_automatic_function_calling_with_parameterized_generic_union_type(
 
 
 @pytest.mark.asyncio
-async def test_google_search_retrieval_async(client):
+async def test_google_search_async(client):
   await client.aio.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents=[
           types.ContentDict(
               {'role': 'user', 'parts': [{'text': 'Why is the sky blue?'}]}
           )
       ],
-      config={'tools': [{'google_search_retrieval': {}}]},
+      config={'tools': [{'google_search': {}}]},
   )
 
 
 def test_empty_tools(client):
   client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='What is the price of GOOG?.',
       config={'tools': []},
   )
@@ -998,9 +1222,9 @@ def test_empty_tools(client):
 
 def test_with_1_empty_tool(client):
   # Bad request for empty tool.
-  with pytest.raises(errors.ClientError):
+  with pytest_helper.exception_if_vertex(client, errors.ClientError):
     client.models.generate_content(
-        model='gemini-1.5-flash',
+        model='gemini-2.5-flash',
         contents='What is the price of GOOG?.',
         config={
             'tools': [{}, get_stock_price],
@@ -1010,11 +1234,11 @@ def test_with_1_empty_tool(client):
 
 
 @pytest.mark.asyncio
-async def test_google_search_retrieval_stream_async(client):
+async def test_google_search_stream_async(client):
   async for part in await client.aio.models.generate_content_stream(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='Why is the sky blue?',
-      config={'tools': [{'google_search_retrieval': {}}]},
+      config={'tools': [{'google_search': {}}]},
   ):
     pass
 
@@ -1023,7 +1247,7 @@ async def test_google_search_retrieval_stream_async(client):
 async def test_vai_search_stream_async(client):
   if client._api_client.vertexai:
     async for part in await client.aio.models.generate_content_stream(
-        model='gemini-1.5-flash',
+        model='gemini-2.5-flash',
         contents='what is vertex ai search?',
         config={
             'tools': [{
@@ -1041,7 +1265,7 @@ async def test_vai_search_stream_async(client):
   else:
     with pytest.raises(ValueError) as e:
       async for part in await client.aio.models.generate_content_stream(
-          model='gemini-1.5-flash',
+          model='gemini-2.5-flash',
           contents='Why is the sky blue?',
           config={
               'tools': [{
@@ -1065,7 +1289,7 @@ def test_automatic_function_calling_with_coroutine_function(client):
 
   with pytest.raises(errors.UnsupportedFunctionError):
     client.models.generate_content(
-        model='gemini-1.5-flash',
+        model='gemini-2.5-flash',
         contents='what is the result of 1000/2?',
         config={
             'tools': [divide_integers],
@@ -1082,7 +1306,7 @@ async def test_automatic_function_calling_with_coroutine_function_async(
     return a // b
 
   response = await client.aio.models.generate_content(
-      model='gemini-2.5-flash-preview-05-20',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -1099,7 +1323,7 @@ async def test_automatic_function_calling_async(client):
     return a // b
 
   response = await client.aio.models.generate_content(
-      model='gemini-2.5-flash-preview-05-20',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -1116,9 +1340,14 @@ async def test_automatic_function_calling_async_with_exception(client):
     return a // b
 
   response = await client.aio.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/0?',
-      config={'tools': [divide_integers]},
+      config={
+          'tools': [divide_integers],
+          'system_instruction': (
+              'you must first look at the tools and then think about answers'
+          ),
+      },
   )
   assert response.automatic_function_calling_history
   assert (
@@ -1131,7 +1360,7 @@ async def test_automatic_function_calling_async_with_exception(client):
 @pytest.mark.asyncio
 async def test_automatic_function_calling_async_float_without_decimal(client):
   response = await client.aio.models.generate_content(
-      model='gemini-2.5-flash-preview-05-20',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000.0/2.0?',
       config={
           'tools': [divide_floats, divide_integers],
@@ -1156,7 +1385,7 @@ async def test_automatic_function_calling_async_with_pydantic_model(client):
       return f'The weather in {city_object.city_name} is sunny and 100 degrees.'
 
   response = await client.aio.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='it is winter now, what is the weather in Boston?',
       config={
           'tools': [get_weather_pydantic_model],
@@ -1177,7 +1406,7 @@ async def test_automatic_function_calling_async_with_async_function(client):
     return 'windy'
 
   response = await client.aio.models.generate_content(
-      model='gemini-2.5-flash-preview-05-20',
+      model='gemini-2.5-flash',
       contents='what is the weather in San Francisco?',
       config={
           'tools': [get_current_weather_async],
@@ -1199,7 +1428,7 @@ async def test_automatic_function_calling_async_with_async_function_stream(
     return 'windy'
 
   response = await client.aio.models.generate_content_stream(
-      model='gemini-2.5-flash-preview-05-20',
+      model='gemini-2.5-flash',
       contents='what is the weather in San Francisco?',
       config={
           'tools': [get_current_weather_async],
@@ -1216,7 +1445,7 @@ async def test_automatic_function_calling_async_with_async_function_stream(
 
 def test_2_function_with_history(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='What is the price of GOOG? And what is the weather in Boston?',
       config={
           'tools': [get_weather, get_stock_price],
@@ -1271,7 +1500,7 @@ def test_2_function_with_history(client):
 @pytest.mark.asyncio
 async def test_2_function_with_history_async(client):
   response = await client.aio.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='What is the price of GOOG? And what is the weather in Boston?',
       config={
           'tools': [get_weather, get_stock_price],
@@ -1353,7 +1582,7 @@ def test_class_method_tools(client):
 
 def test_disable_afc_in_any_mode(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config=types.GenerateContentConfig(
           tools=[divide_integers],
@@ -1369,7 +1598,7 @@ def test_disable_afc_in_any_mode(client):
 
 def test_afc_once_in_any_mode(client):
   response = client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config=types.GenerateContentConfig(
           tools=[divide_integers],
@@ -1405,7 +1634,7 @@ def test_code_execution_tool(client):
 def test_afc_logs_to_logger_instance(client, caplog):
   caplog.set_level(logging.DEBUG, logger='google_genai.models')
   client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -1430,7 +1659,7 @@ def test_suppress_logs_with_sdk_logger(client, caplog):
   sdk_logger = logging.getLogger('google_genai.models')
   sdk_logger.setLevel(logging.ERROR)
   client.models.generate_content(
-      model='gemini-1.5-flash',
+      model='gemini-2.5-flash',
       contents='what is the result of 1000/2?',
       config={
           'tools': [divide_integers],
@@ -1454,7 +1683,7 @@ def test_tools_chat_curation(client, caplog):
   }
 
   chat = client.chats.create(
-      model='gemini-2.5-flash-preview-05-20',
+      model='gemini-2.5-flash',
       config=config,
   )
 
@@ -1468,3 +1697,65 @@ def test_tools_chat_curation(client, caplog):
 
   history = chat.get_history(curated=True)
   assert len(history) == 4
+
+
+def test_function_declaration_with_callable(client):
+  response = client.models.generate_content(
+      model='gemini-2.5-pro',
+      contents=(
+          'Divide 1000 by 2. And tell'
+          ' me the weather in London.'
+      ),
+      config={
+          'tools': [
+              divide_integers,
+              {'function_declarations': function_declarations},
+          ],
+      },
+  )
+  assert response.function_calls is not None
+
+def test_function_declaration_with_callable_stream_now(client):
+  for chunk in client.models.generate_content_stream(
+      model='gemini-2.5-pro',
+      contents='Divide 1000 by 2. And tell me the weather in London.',
+      config={
+          'tools': [
+              divide_integers,
+              {'function_declarations': function_declarations},
+          ],
+      },
+  ):
+    pass
+
+@pytest.mark.asyncio
+async def test_function_declaration_with_callable_async(client):
+  response = await client.aio.models.generate_content(
+      model='gemini-2.5-pro',
+      contents=(
+          'Divide 1000 by 2. And tell'
+          ' me the weather in London.'
+      ),
+      config={
+          'tools': [
+              divide_integers,
+              {'function_declarations': function_declarations},
+          ],
+      },
+  )
+  assert response.function_calls is not None
+
+
+@pytest.mark.asyncio
+async def test_function_declaration_with_callable_async_stream(client):
+    async for chunk in await client.aio.models.generate_content_stream(
+        model='gemini-2.5-pro',
+        contents='Divide 1000 by 2. And tell me the weather in London.',
+        config={
+            'tools': [
+                divide_integers,
+                {'function_declarations': function_declarations},
+            ],
+        },
+    ):
+      pass
